detects array from bigiq and displays accordingly
